requirements.txt
altair==5.2.0
archspec==0.2.1
attrs==23.2.0
blinker==1.7.0
boltons==23.0.0
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
cryptography==41.0.7
distro==1.8.0
ffmpeg-python==0.2.0
filelock==3.13.1
fsspec==2024.3.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
idna==3.4
Jinja2==3.1.3
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.1
numba==0.59.0
numpy==1.26.4
openai-whisper==20231117
packaging==23.1
pandas==2.2.1
pillow==10.2.0
pip==23.3.1
platformdirs==3.10.0
pluggy==1.0.0
protobuf==4.25.3
pyarrow==15.0.1
pycosat==0.6.6
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml==0.17.21
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.65.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
wheel==0.41.2
zstandard==0.19.0
